moving all of the autocleaning and jlisp interpreter config into separate module
