+++ buckaroo/dataflow/autocleaning.py
@@ -1,0 +1,91 @@
+import pandas as pd
+from buckaroo.pluggable_analysis_framework.utils import (filter_analysis)
+from ..customizations.all_transforms import configure_buckaroo, DefaultCommandKlsList
+
+def dumb_merge_ops(existing_ops, cleaning_ops):
+    """ strip cleaning_ops from existing_ops, reinsert cleaning_ops at the beginning """
+    a = existing_ops.copy()
+    a.extend(cleaning_ops)
+    return a
+
+
+SENTINEL_DF_1 = pd.DataFrame({'foo'  :[10, 20], 'bar' : ["asdf", "iii"]})
+SENTINEL_DF_2 = pd.DataFrame({'col1' :[55, 55], 'col2': ["pppp", "333"]})
+SENTINEL_DF_3 = pd.DataFrame({'pp'   :[99, 33], 'ee':   [     6,     9]})
+SENTINEL_DF_4 = pd.DataFrame({'vvv'  :[12, 49], 'oo':   [ 'ccc', 'www']})
+
+
+
+class SentinelAutocleaning:
+
+    def handle_ops_and_clean(self, df, cleaning_method, existing_operations):
+        cleaning_ops = []
+        generated_code = ""
+        if cleaning_method == "one op":
+            cleaning_ops =  [""]
+            generated_code = "codegen 1"
+            generated_code = "codegen 2"
+        elif cleaning_method == "two op":
+            cleaning_ops = ["", ""]
+
+        merged_operations = dumb_merge_ops(existing_operations, cleaning_ops)
+        
+        if len(merged_operations) == 1:
+            cleaned_df = SENTINEL_DF_1
+        elif len(merged_operations) == 2:
+            cleaned_df = SENTINEL_DF_2
+        else:
+            cleaned_df = df
+        return [cleaned_df, {}, generated_code, merged_operations]
+
+def merge_ops(existing_ops, cleaning_ops):
+    """ strip cleaning_ops from existing_ops, reinsert cleaning_ops at the beginning """
+    pass
+
+class Autocleaning:
+
+
+    command_klasses = [DefaultCommandKlsList]
+
+    autocleaning_analysis_klasses = []
+    
+
+    ### start code interpreter block
+    def _setup_from_command_kls_list(self):
+
+
+        #used to initially setup the interpreter, and when a command
+        #is added interactively
+        c_klasses = self.command_klasses
+        c_defaults, c_patterns, df_interpreter, gencode_interpreter = configure_buckaroo(c_klasses)
+        self.df_interpreter, self.gencode_interpreter = df_interpreter, gencode_interpreter
+        self.commandConfig = dict(argspecs=c_patterns, defaultArgs=c_defaults)
+        self.autocleaning_genops = filter_analysis(self.autocleaning_analysis_klasses, "autocleaning_ops")
+
+    def add_command(self, incomingCommandKls):
+        without_incoming = [x for x in self.command_classes if not x.__name__ == incomingCommandKls.__name__]
+        without_incoming.append(incomingCommandKls)
+        self.command_klasses = without_incoming
+        self.setup_from_command_kls_list()
+
+    def _run_df_interpreter(self, df, operations):
+        full_ops = [{'symbol': 'begin'}]
+        full_ops.extend(operations)
+        if len(full_ops) == 1:
+            return df
+        return self.buckaroo_transform(full_ops , df)
+
+    def run_code_generator(self, operations):
+        if len(operations) == 0:
+            return 'no operations'
+        return self.gencode_interpreter(operations)
+
+    def handle_ops_and_clean_orig(self, df, cleaning_method, existing_operations):
+        if self.sampled_df is None:
+            return None
+        cleaning_operations, cleaning_sd = self._run_cleaning(df, cleaning_method)
+        merged_operations = merge_ops(existing_operations, cleaning_operations)
+        cleaned_df = self._run_df_interpreter(df, merged_operations)
+        generated_code = self._run_code_generator(merged_operations)
+        self.cleaned = [cleaned_df, cleaning_sd, generated_code, merged_operations]
+
